pyproject.toml
[project]
name = "wind-farm-radar-evaluation"
version = "0.1.0"
description = "风电场对海岸监视雷达影响评估系统"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
    "scikit-learn>=1.0.0",
    "PyYAML>=6.0",
    "radarsimpy>=0.5.0",
    "pymeshlab>=0.2.0",
    "reportlab>=3.6.0",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.910",
    "jupyter>=1.0",
    "ipython>=7.0",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/wind-farm-radar-evaluation"
Repository = "https://github.com/yourusername/wind-farm-radar-evaluation"
Documentation = "https://yourusername.github.io/wind-farm-radar-evaluation"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wind_farm_radar_evaluation"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--verbose --cov=src/wind_farm_radar_evaluation --cov-report=html"
